PID kontroller ready

+++ node_2/servo_driver.h
@@ -1,8 +1,8 @@
 #ifndef SERVO_H
 #define SERVO_H
+#include <stdint.h>
 
-
-void servo_set_angle(int pos_x);
+void servo_set_angle(int8_t pos_x);
